modal and got rid of allworkouts
